requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==25.1.0
autobahn==23.1.2
Automat==24.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
billiard==4.1.0
celery==5.3.4
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.19
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.3.1
django-storages==1.14.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
kombu==5.3.2
MarkupPy==1.14
msgpack==1.1.0
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==25.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.32.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.9
xlrd==2.0.1
xlwt==1.3.0
zope.interface==7.2
Pillow==10.4.0
